--- src/infrastructure/storage/src/test/kotlin/yapp/be/storage/test/StorageTest.kt
@@ -1,12 +1,0 @@
-package yapp.be.storage.test
-
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
-import support.container.ContainerTestBase
-
-class StorageTest : ContainerTestBase() {
-    @Test
-    @DisplayName("DDL과 Entity 일치여부 validation 테스트")
-    fun contextLoads() {
-    }
-}
