+++ src/infrastructure/storage/src/testFixtures/kotlin/support/container/MySqlContainerProperties.kt
@@ -1,6 +1,6 @@
 package support.container
 
-object ContainerProperties {
+object MySqlContainerProperties {
     const val imageName = "mysql:8.0"
     const val PORT = 3306
     const val PASSWORD = "1q2w3e4r"
